fetch package refactored
