+++ src/main/java/com/configcat/ConfigCatClient.java
@@ -5,6 +5,7 @@ import com.configcat.cache.NullConfigCache;
 import com.configcat.evaluation.EvaluationDetails;
 import com.configcat.evaluation.EvaluationResult;
 import com.configcat.evaluation.RolloutEvaluator;
+import com.configcat.fetch.ConfigFetcher;
 import com.configcat.hooks.ConfigCatHooks;
 import com.configcat.log.ConfigCatLogMessages;
 import com.configcat.log.ConfigCatLogger;
