+++ src/main/java/com/configcat/Result.java
@@ -9,19 +9,19 @@ public final class Result<T> {
         this.error = error;
     }
 
-    T value() {
+    public T value() {
         return this.value;
     }
 
-    String error() {
+    public String error() {
         return this.error;
     }
 
-    static <T> Result<T> error(String error, T value) {
+    public static <T> Result<T> error(String error, T value) {
         return new Result<>(value, error);
     }
 
-    static <T> Result<T> success(T value) {
+    public static <T> Result<T> success(T value) {
         return new Result<>(value, null);
     }
 }
