+++ src/main/java/com/configcat/fetch/ConfigFetcher.java
@@ -1,5 +1,9 @@
-package com.configcat;
+package com.configcat.fetch;
 
+import com.configcat.Constants;
+import com.configcat.DateTimeUtils;
+import com.configcat.Result;
+import com.configcat.Utils;
 import com.configcat.log.ConfigCatLogMessages;
 import com.configcat.log.ConfigCatLogger;
 import com.configcat.models.Config;
@@ -13,69 +17,7 @@ import java.io.IOException;
 import java.net.SocketTimeoutException;
 import java.util.concurrent.atomic.AtomicBoolean;
 
-class FetchResponse {
-    public enum Status {
-        FETCHED,
-        NOT_MODIFIED,
-        FAILED
-    }
-
-    private final Status status;
-    private final Entry entry;
-    private final String error;
-    private final boolean fetchTimeUpdatable;
-    private final String fetchTime;
-
-    public boolean isFetched() {
-        return this.status == Status.FETCHED;
-    }
-
-    public boolean isNotModified() {
-        return this.status == Status.NOT_MODIFIED;
-    }
-
-    public boolean isFailed() {
-        return this.status == Status.FAILED;
-    }
-
-    public boolean isFetchTimeUpdatable() {
-        return fetchTimeUpdatable;
-    }
-
-    public String getFetchTime() {
-        return this.fetchTime;
-    }
-
-    public Entry entry() {
-        return this.entry;
-    }
-
-    public String error() {
-        return this.error;
-    }
-
-    FetchResponse(Status status, Entry entry, String error, boolean fetchTimeUpdatable, String fetchTime) {
-        this.status = status;
-        this.entry = entry;
-        this.error = error;
-        this.fetchTimeUpdatable = fetchTimeUpdatable;
-        this.fetchTime = fetchTime;
-    }
-
-    public static FetchResponse fetched(Entry entry, String fetchTime) {
-        return new FetchResponse(Status.FETCHED, entry == null ? Entry.EMPTY : entry, null, false, fetchTime);
-    }
-
-    public static FetchResponse notModified(String fetchTime) {
-        return new FetchResponse(Status.NOT_MODIFIED, Entry.EMPTY, null, true, fetchTime);
-    }
-
-    public static FetchResponse failed(String error, boolean fetchTimeUpdatable, String fetchTime) {
-        return new FetchResponse(Status.FAILED, Entry.EMPTY, error, fetchTimeUpdatable, fetchTime);
-    }
-}
-
-class ConfigFetcher implements Closeable {
+public class ConfigFetcher implements Closeable {
     private final AtomicBoolean closed = new AtomicBoolean(false);
     private final ConfigCatLogger logger;
     private final OkHttpClient httpClient;
@@ -91,7 +33,7 @@ class ConfigFetcher implements Closeable {
         FORCE_REDIRECT
     }
 
-    ConfigFetcher(OkHttpClient httpClient,
+    public ConfigFetcher(OkHttpClient httpClient,
                   ConfigCatLogger logger,
                   String sdkKey,
                   String url,
