+++ src/test/java/com/configcat/ConfigFetcherTest.java
@@ -2,6 +2,8 @@ package com.configcat;
 
 import com.configcat.cache.CacheException;
 import com.configcat.cache.ConfigCache;
+import com.configcat.fetch.ConfigFetcher;
+import com.configcat.fetch.FetchResponse;
 import com.configcat.hooks.ConfigCatHooks;
 import com.configcat.log.ConfigCatLogger;
 import com.configcat.log.LogLevel;
