+++ src/test/java/com/configcat/ManualPollingPolicyTest.java
@@ -2,6 +2,7 @@ package com.configcat;
 
 import com.configcat.cache.InMemoryCache;
 import com.configcat.cache.NullConfigCache;
+import com.configcat.fetch.ConfigFetcher;
 import com.configcat.hooks.ConfigCatHooks;
 import com.configcat.log.ConfigCatLogger;
 import com.configcat.polling.PollingMode;
